+++ backend/Dockerfile
@@ -7,7 +7,6 @@ WORKDIR /backend
 COPY requirements.txt /backend
 RUN pip install -r requirements.txt
 
-COPY app.py /backend
-COPY config.py /backend
-COPY lists.py /backend
-CMD python app.py
+COPY run.py /backend
+COPY washroomcatalog /backend/washroomcatalog
+CMD python run.py
